cluster

+++ deploy/server.py
@@ -194,7 +194,7 @@ def dist_2set(group1, group2):
         i_g2 = [np.sum(np.square(j, i)) for j in group2]
         dist.append(min(i_g2))
 
-    return int(min(dist)), float(np.average(dist)), int(max(dist))
+    return float(min(dist)), float(np.average(dist)), float(max(dist))
 
 
 if __name__ == '__main__':
